requirements.txt
transformers==4.23.1
torch==1.12.1
Pillow==9.3.0
tokenizers==0.13.1
Flask